requirements.txt
numpy
opencv-python-headless
Pillow
transformers
scipy
tqdm

openai>=1.0.0
google-generativeai
deep-translator
python-dotenv

fastapi
uvicorn
pydantic>=2.0.0
python-multipart
requests
